# stl-verify/ts/ui/src/generated/openapi-schema.json
{
  "components": {
    "schemas": {
...
      "ExposureBucketResponse": {
        "description": "Priced receipt-token exposure within a single time bucket (LOCF gap-filled).",
        "properties": {
          "bucket_start": {
            "description": "Inclusive start of the time bucket (UTC).",
            "format": "date-time",
            "title": "Bucket Start",
            "type": "string"
          },
          "exposure_usd": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Sum across the prime's receipt-token positions of the carried-forward balance valued at the latest underlying oracle price (USD), serialized as a JSON string. `null` for leading buckets before the first observation.",
            "examples": [
              "1459014561.88"
            ],
            "title": "Exposure Usd"
          }
        },
        "required": [
          "bucket_start"
        ],
        "title": "ExposureBucketResponse",
        "type": "object"
      },
      "ExposureEnvelope": {
        "description": "Per-prime exposure time series, gap-filled into buckets.",
        "properties": {
          "data": {
            "description": "Priced exposure per time bucket.",
            "items": {
              "$ref": "#/components/schemas/ExposureBucketResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "mode": {
            "const": "aggregated",
            "description": "Always `aggregated`: a gap-filled time series.",
            "title": "Mode",
            "type": "string"
          },
          "window": {
            "$ref": "#/components/schemas/TimeSeriesWindow",
            "description": "The window and resolution applied to this response."
          }
        },
        "required": [
          "mode",
          "window",
          "data"
        ],
        "title": "ExposureEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PrimeDebtBucketResponse": {
        "description": "Last observed debt within a single time bucket (LOCF gap-filled).",
        "properties": {
          "bucket_start": {
            "description": "Inclusive start of the time bucket (UTC).",
            "format": "date-time",
            "title": "Bucket Start",
            "type": "string"
          },
          "debt_wad": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Last observed debt in `wad` units carried forward into the bucket, serialized as a JSON string. `null` for leading buckets before the first observation.",
            "examples": [
              "1234567890000000000000"
            ],
            "title": "Debt Wad"
          }
        },
        "required": [
          "bucket_start"
        ],
        "title": "PrimeDebtBucketResponse",
        "type": "object"
      },
      "PrimeDebtEnvelope": {
        "description": "Prime debt response: raw snapshots or aggregated time buckets.",
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PrimeDebtSnapshotResponse"
                },
                "type": "array"
              },
              {
                "items": {
                  "$ref": "#/components/schemas/PrimeDebtBucketResponse"
                },
                "type": "array"
              }
            ],
            "description": "Snapshots when `mode=raw`, value buckets when `mode=aggregated`.",
            "title": "Data"
          },
          "mode": {
            "description": "`raw` for snapshots, `aggregated` for time buckets.",
            "enum": [
              "raw",
              "aggregated"
            ],
            "title": "Mode",
            "type": "string"
          },
          "window": {
            "$ref": "#/components/schemas/TimeSeriesWindow",
            "description": "The window and resolution applied to this response."
          }
        },
        "required": [
          "mode",
          "window",
          "data"
        ],
        "title": "PrimeDebtEnvelope",
        "type": "object"
      },
      "PrimeDebtSnapshotResponse": {
        "description": "A single observed prime-debt position at a point in time.",
        "example": {
          "block_number": 18000000,
          "block_version": 1,
          "debt_wad": "1234567890000000000000",
          "ilk_name": "ALLOCATOR-NEXUS-A",
          "prime_address": "0x1234567890abcdef1234567890abcdef12345678",
          "prime_name": "Acme Prime",
          "synced_at": "2026-05-07T12:00:00Z"
        },
        "properties": {
          "block_number": {
            "description": "Block number the snapshot was observed at.",
            "examples": [
              18000000
            ],
            "title": "Block Number",
            "type": "integer"
          },
          "block_version": {
            "description": "Cache-key version that increments on chain reorgs.",
            "examples": [
              1
            ],
            "title": "Block Version",
            "type": "integer"
          },
          "debt_wad": {
            "description": "Outstanding debt in MakerDAO `wad` units (1e18 fixed-point). Decimal serialized as a JSON string to preserve precision.",
            "examples": [
              "1234567890000000000000"
            ],
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Debt Wad",
            "type": "string"
          },
          "ilk_name": {
            "description": "Maker `ilk` (collateral type) the debt is denominated against.",
            "examples": [
              "ALLOCATOR-NEXUS-A"
            ],
            "title": "Ilk Name",
            "type": "string"
          },
          "prime_address": {
            "description": "Prime's 0x-prefixed Ethereum address.",
            "examples": [
              "0x1234567890abcdef1234567890abcdef12345678"
            ],
            "title": "Prime Address",
            "type": "string"
          },
          "prime_name": {
            "description": "Human-readable prime name.",
            "examples": [
              "Acme Prime"
            ],
            "title": "Prime Name",
            "type": "string"
          },
          "synced_at": {
            "description": "Server-side time the snapshot was persisted.",
            "format": "date-time",
            "title": "Synced At",
            "type": "string"
          }
        },
        "required": [
          "prime_address",
          "prime_name",
          "ilk_name",
          "debt_wad",
          "block_number",
          "block_version",
          "synced_at"
        ],
        "title": "PrimeDebtSnapshotResponse",
        "type": "object"
      },
...
      "TimeSeriesResolution": {
        "description": "Allowed ISO-8601 durations for time-series downsampling.",
        "enum": [
          "PT1M",
          "PT5M",
          "PT15M",
          "PT1H",
          "PT6H",
          "P1D"
        ],
        "title": "TimeSeriesResolution",
        "type": "string"
      },
      "TimeSeriesWindow": {
        "description": "The resolved window and resolution actually applied to a request.\n\nEchoing this back lets consumers distinguish an empty result caused by the\nwindow from one caused by the absence of data.",
        "properties": {
          "from_timestamp": {
            "description": "Inclusive lower bound applied (UTC).",
            "format": "date-time",
            "title": "From Timestamp",
            "type": "string"
          },
          "interval_ms": {
            "description": "Resolution width in milliseconds.",
            "title": "Interval Ms",
            "type": "integer"
          },
          "resolution": {
            "$ref": "#/components/schemas/TimeSeriesResolution",
            "description": "Resolution applied (relevant when aggregated)."
          },
          "to_timestamp": {
            "description": "Inclusive upper bound applied (UTC).",
            "format": "date-time",
            "title": "To Timestamp",
            "type": "string"
          }
        },
        "required": [
          "from_timestamp",
          "to_timestamp",
          "resolution",
          "interval_ms"
        ],
        "title": "TimeSeriesWindow",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Verify-side HTTP API for the STL pipeline.\n\nEndpoints expose primes (capital allocators), their allocations and debt, decoded protocol events, token catalog and pricing, and risk-capital computations (RRC, bad debt, breakdown).",
    "title": "stl-verify",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/primes/{prime_id}/debt": {
      "get": {
        "description": "Return debt snapshots for a prime, newest first, inside a `{mode, window, data}` envelope. Results are time-windowed (default last 24h). Returns `404` if the prime is unknown. Each snapshot carries the `block_number`/`block_version` it was observed at; consumers can use `block_version` to detect reorg-driven re-emissions. Set `aggregate=true` for the last debt value per time bucket (gap-filled).",
        "operationId": "list_prime_debt_snapshots_v1_primes__prime_id__debt_get",
        "parameters": [
          {
            "description": "A prime's 0x-prefixed ALM **proxy** address on one chain \u2014 not a prime identifier. A prime allocates through one proxy per chain; list them via `GET /v1/primes` and group by `prime_vault_address`.",
            "in": "path",
            "name": "prime_id",
            "required": true,
            "schema": {
              "description": "A prime's 0x-prefixed ALM **proxy** address on one chain \u2014 not a prime identifier. A prime allocates through one proxy per chain; list them via `GET /v1/primes` and group by `prime_vault_address`.",
              "title": "Prime Id",
              "type": "string"
            }
          },
          {
            "description": "Max snapshots returned (default 100, max 500).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "Max snapshots returned (default 100, max 500).",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Inclusive lower timestamp bound (ISO-8601). Defaults to 24h before `to_timestamp`.",
            "in": "query",
            "name": "from_timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive lower timestamp bound (ISO-8601). Defaults to 24h before `to_timestamp`.",
              "title": "From Timestamp"
            }
          },
          {
            "description": "Inclusive upper timestamp bound (ISO-8601). Defaults to the current UTC time.",
            "in": "query",
            "name": "to_timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive upper timestamp bound (ISO-8601). Defaults to the current UTC time.",
              "title": "To Timestamp"
            }
          },
          {
            "description": "ISO-8601 duration resolution (for example `PT5M`, `PT1H`). Used for time-bucketing when `aggregate=true`; defaults to the finest resolution allowed for the window.",
            "in": "query",
            "name": "resolution",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TimeSeriesResolution"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 duration resolution (for example `PT5M`, `PT1H`). Used for time-bucketing when `aggregate=true`; defaults to the finest resolution allowed for the window.",
              "title": "Resolution"
            }
          },
          {
            "description": "When true, return time-bucketed aggregates instead of raw rows.",
            "in": "query",
            "name": "aggregate",
            "required": false,
            "schema": {
              "default": false,
              "description": "When true, return time-bucketed aggregates instead of raw rows.",
              "title": "Aggregate",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrimeDebtEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List prime debt snapshots",
        "tags": [
          "primes"
        ]
      }
    },
    "/v1/primes/{prime_id}/exposure": {
      "get": {
        "description": "Return the prime's priced receipt-token exposure over time, gap-filled (LOCF) into buckets. Per bucket, each receipt-token position's carried-forward balance is valued at the latest underlying oracle price and summed (the current `balance * price` exposure extended over time). Direct (non-receipt-token) holdings are excluded, matching the risk-capital exposure basis. Returns `404` if the prime is unknown. Defaults to the last 24h; pass a window and `resolution` for longer ranges.",
        "operationId": "list_prime_exposure_v1_primes__prime_id__exposure_get",
        "parameters": [
          {
            "description": "A prime's 0x-prefixed ALM **proxy** address on one chain \u2014 not a prime identifier. A prime allocates through one proxy per chain; list them via `GET /v1/primes` and group by `prime_vault_address`.",
            "in": "path",
            "name": "prime_id",
            "required": true,
            "schema": {
              "description": "A prime's 0x-prefixed ALM **proxy** address on one chain \u2014 not a prime identifier. A prime allocates through one proxy per chain; list them via `GET /v1/primes` and group by `prime_vault_address`.",
              "title": "Prime Id",
              "type": "string"
            }
          },
          {
            "description": "Max buckets returned (default 100, max 500).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "Max buckets returned (default 100, max 500).",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Inclusive lower timestamp bound (ISO-8601). Defaults to 24h before `to_timestamp`.",
            "in": "query",
            "name": "from_timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive lower timestamp bound (ISO-8601). Defaults to 24h before `to_timestamp`.",
              "title": "From Timestamp"
            }
          },
          {
            "description": "Inclusive upper timestamp bound (ISO-8601). Defaults to the current UTC time.",
            "in": "query",
            "name": "to_timestamp",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive upper timestamp bound (ISO-8601). Defaults to the current UTC time.",
              "title": "To Timestamp"
            }
          },
          {
            "description": "ISO-8601 duration resolution (for example `PT5M`, `PT1H`). Used for time-bucketing when `aggregate=true`; defaults to the finest resolution allowed for the window.",
            "in": "query",
            "name": "resolution",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TimeSeriesResolution"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 duration resolution (for example `PT5M`, `PT1H`). Used for time-bucketing when `aggregate=true`; defaults to the finest resolution allowed for the window.",
              "title": "Resolution"
            }
          },
          {
            "description": "When true, return time-bucketed aggregates instead of raw rows.",
            "in": "query",
            "name": "aggregate",
            "required": false,
            "schema": {
              "default": false,
              "description": "When true, return time-bucketed aggregates instead of raw rows.",
              "title": "Aggregate",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExposureEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prime exposure time series",
        "tags": [
          "primes",
          "capital"
        ]
      }
    },
...
  },
...
}